+++ fastly/model_logging_logentries_additional.go
@@ -23,7 +23,7 @@ type LoggingLogentriesAdditional struct {
 	Port *int32 `json:"port,omitempty"`
 	// Use token based authentication.
 	Token *string `json:"token,omitempty"`
-	UseTLS *LoggingUseTLS `json:"use_tls,omitempty"`
+	UseTLS *LoggingUseTLSString `json:"use_tls,omitempty"`
 	// The region to which to stream logs.
 	Region *string `json:"region,omitempty"`
 	AdditionalProperties map[string]any
@@ -39,7 +39,7 @@ func NewLoggingLogentriesAdditional() *LoggingLogentriesAdditional {
 	this := LoggingLogentriesAdditional{}
 	var port int32 = 20000
 	this.Port = &port
-	var useTLS LoggingUseTLS = LOGGINGUSETLS_no_tls
+	var useTLS LoggingUseTLSString = LOGGINGUSETLSSTRING_no_tls
 	this.UseTLS = &useTLS
 	return &this
 }
@@ -51,7 +51,7 @@ func NewLoggingLogentriesAdditionalWithDefaults() *LoggingLogentriesAdditional {
 	this := LoggingLogentriesAdditional{}
 	var port int32 = 20000
 	this.Port = &port
-	var useTLS LoggingUseTLS = LOGGINGUSETLS_no_tls
+	var useTLS LoggingUseTLSString = LOGGINGUSETLSSTRING_no_tls
 	this.UseTLS = &useTLS
 	return &this
 }
@@ -121,9 +121,9 @@ func (o *LoggingLogentriesAdditional) SetToken(v string) {
 }
 
 // GetUseTLS returns the UseTLS field value if set, zero value otherwise.
-func (o *LoggingLogentriesAdditional) GetUseTLS() LoggingUseTLS {
+func (o *LoggingLogentriesAdditional) GetUseTLS() LoggingUseTLSString {
 	if o == nil || o.UseTLS == nil {
-		var ret LoggingUseTLS
+		var ret LoggingUseTLSString
 		return ret
 	}
 	return *o.UseTLS
@@ -131,7 +131,7 @@ func (o *LoggingLogentriesAdditional) GetUseTLS() LoggingUseTLS {
 
 // GetUseTLSOk returns a tuple with the UseTLS field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *LoggingLogentriesAdditional) GetUseTLSOk() (*LoggingUseTLS, bool) {
+func (o *LoggingLogentriesAdditional) GetUseTLSOk() (*LoggingUseTLSString, bool) {
 	if o == nil || o.UseTLS == nil {
 		return nil, false
 	}
@@ -147,8 +147,8 @@ func (o *LoggingLogentriesAdditional) HasUseTLS() bool {
 	return false
 }
 
-// SetUseTLS gets a reference to the given LoggingUseTLS and assigns it to the UseTLS field.
-func (o *LoggingLogentriesAdditional) SetUseTLS(v LoggingUseTLS) {
+// SetUseTLS gets a reference to the given LoggingUseTLSString and assigns it to the UseTLS field.
+func (o *LoggingLogentriesAdditional) SetUseTLS(v LoggingUseTLSString) {
 	o.UseTLS = &v
 }