+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [v1.0.0-beta.31](https://github.com/fastly/fastly-go/releases/tag/release/v1.0.0-beta.31) (2024-07-01)
+
+**Bug fixes:**
+
+- fix(logging): For several endpoints, correct use_tls to be string
+
+**Enhancements:**
+
+- feat(content): `/content/edge_check` endpoint now returns informational values in `hash` when a timeout occurs or when
+  an object is too large.
+- feat(logging-datalog): Added additional regions
+
 ## [vv1.0.0-beta.30](https://github.com/fastly/fastly-go/releases/tag/release/vv1.0.0-beta.30) (2024-06-27)
 
 **Bug fixes:**